Implement controls
Player now moves according to events from the keyboard.

Future implementation may add mouse-following and other event-actions. InputManager is still a work in progress for this implementation.

+++ MoveableThing.h
@@ -10,7 +10,7 @@ public:
 	MoveableThing(sf::RenderWindow& win);
 	~MoveableThing();
 	virtual void takeTurn(void);
-	bool move(int deltaX, int deltaY);
+	bool move(float deltaX, float deltaY);
 	bool resolveCollisions(int x, int y);
 
 // protected:
